Auditing Core Data layer - audited/tested Award, AwardRecipient, District, DistrictEventPoints, DistrictRanking. Bumped TBAKit without fixing caching yet

+++ tba-unit-tests/Extensions/Int16+Suffix_Tests.swift
@@ -13,21 +13,21 @@ class Int16SuffixTestCase: XCTestCase {
 
     func test_suffix_multiple(multiplier: Int) {
         // A weird one, but technically right
-        XCTAssertEqual(Int16(0 * multiplier).suffix(), "th")
+        XCTAssertEqual((0 * multiplier).suffix, "th")
         // Test lastOne + a few
-        XCTAssertEqual(Int16(1 * multiplier).suffix(), "st")
-        XCTAssertEqual(Int16(2 * multiplier).suffix(), "nd")
-        XCTAssertEqual(Int16(3 * multiplier).suffix(), "rd")
-        XCTAssertEqual(Int16(4 * multiplier).suffix(), "th")
-        XCTAssertEqual(Int16(5 * multiplier).suffix(), "th")
+        XCTAssertEqual((1 * multiplier).suffix, "st")
+        XCTAssertEqual((2 * multiplier).suffix, "nd")
+        XCTAssertEqual((3 * multiplier).suffix, "rd")
+        XCTAssertEqual((4 * multiplier).suffix, "th")
+        XCTAssertEqual((5 * multiplier).suffix, "th")
         // Test lastTwo range inclusive
-        XCTAssertEqual(Int16(10 * multiplier).suffix(), "th")
-        XCTAssertEqual(Int16(11 * multiplier).suffix(), "th")
-        XCTAssertEqual(Int16(12 * multiplier).suffix(), "th")
-        XCTAssertEqual(Int16(20 * multiplier).suffix(), "th")
-        XCTAssertEqual(Int16(21 * multiplier).suffix(), "st")
+        XCTAssertEqual((10 * multiplier).suffix, "th")
+        XCTAssertEqual((11 * multiplier).suffix, "th")
+        XCTAssertEqual((12 * multiplier).suffix, "th")
+        XCTAssertEqual((20 * multiplier).suffix, "th")
+        XCTAssertEqual((21 * multiplier).suffix, "st")
         // Test a big number
-        XCTAssertEqual(Int16(32007 * multiplier).suffix(), "th")
+        XCTAssertEqual((32007 * multiplier).suffix, "th")
     }
 
 }
